# Use Map of collections instead of collection to speed up access by path