--- a/src/main/java/net/datafaker/service/FakeValues.java
+++ b/src/main/java/net/datafaker/service/FakeValues.java
@@ -95,4 +95,8 @@ public class FakeValues implements FakeValuesInterface {
     boolean supportsPath(String path) {
         return this.path.equals(path);
     }
+
+    public String getPath() {
+        return path;
+    }
 }