--- a/src/main/java/net/datafaker/service/FakeValuesGrouping.java
+++ b/src/main/java/net/datafaker/service/FakeValuesGrouping.java
@@ -1,29 +1,30 @@
 package net.datafaker.service;
 
 import java.util.ArrayList;
-import java.util.List;
+import java.util.Collection;
+import java.util.Collections;
+import java.util.HashMap;
 import java.util.Map;
 
 public class FakeValuesGrouping implements FakeValuesInterface {
 
-    private final List<FakeValues> fakeValuesList = new ArrayList<>();
+    private final Map<String, Collection<FakeValues>> fakeValues = new HashMap<>();
 
-    public void add(FakeValues fakeValues) {
-        fakeValuesList.add(fakeValues);
+    public void add(FakeValues fakeValue) {
+        fakeValues.putIfAbsent(fakeValue.getPath(), new ArrayList<>());
+        fakeValues.get(fakeValue.getPath()).add(fakeValue);
     }
 
     @Override
     @SuppressWarnings({"unchecked", "rawtypes"})
     public Map get(String key) {
         Map result = null;
-        for (FakeValues fakeValues : fakeValuesList) {
-            if (fakeValues.supportsPath(key)) {
-                if (result != null) {
-                    final Map newResult = fakeValues.get(key);
-                    result.putAll(newResult);
-                } else {
-                    result = fakeValues.get(key);
-                }
+        for (FakeValues fakeValues : fakeValues.getOrDefault(key, Collections.emptyList())) {
+            if (result != null) {
+                final Map newResult = fakeValues.get(key);
+                result.putAll(newResult);
+            } else {
+                result = fakeValues.get(key);
             }
         }
         return result;