Another PR to speed up `faker.expression` a bit. 
The idea is to use map to access fakevalues in `FakeValuesGrouping` which is O(1) instead of collection which in O(n)

Some measurements shows about 20-30% improvement
before
```
Benchmark                            Mode  Cnt    Score    Error   Units
JmhTest._bothifyExpression          thrpt    5  533.255 ± 22.638  ops/ms
JmhTest._letterifyExpression        thrpt    5  531.161 ± 12.172  ops/ms
JmhTest._optionsExpression          thrpt    5  370.678 ± 16.982  ops/ms
JmhTest._regexifyExpression         thrpt    5  238.741 ± 15.081  ops/ms
JmhTest.oneMethodCallExpression     thrpt    5  337.323 ±  1.139  ops/ms
JmhTest.threeMethodsCallExpression  thrpt    5   67.555 ±  6.570  ops/ms
```
after
```
Benchmark                            Mode  Cnt    Score   Error   Units
JmhTest._bothifyExpression          thrpt    5  617.794 ± 5.524  ops/ms
JmhTest._letterifyExpression        thrpt    5  626.325 ± 7.130  ops/ms
JmhTest._optionsExpression          thrpt    5  438.888 ± 9.530  ops/ms
JmhTest._regexifyExpression         thrpt    5  256.661 ± 8.576  ops/ms
JmhTest.oneMethodCallExpression     thrpt    5  405.579 ± 2.753  ops/ms
JmhTest.threeMethodsCallExpression  thrpt    5  106.244 ± 1.232  ops/ms
```

code for benchmark is still same as at https://github.com/datafaker-net/datafaker/pull/25